+++ plugins/nf-schema/src/main/nextflow/validation/utils/Common.groovy
@@ -59,8 +59,17 @@ public class Common {
     //
     // Find a value in a nested map
     //
-    public static Object findDeep(Map m, String key) {
-        if (m.containsKey(key)) return m[key]
-        m.findResult { k, v -> v instanceof Map ? findDeep(v, key) : null }
+    public static Object findDeep(Object m, String key) {
+        if (m instanceof Map) {
+            if (m.containsKey(key)) {
+                return m[key]
+            } else {
+                return m.findResult { k, v -> findDeep(v, key) }
+            }
+        } 
+        else if (m instanceof List) {
+            return m.findResult { element -> findDeep(element, key) }
+        }
+        return null
     }
 }